Zen is now installed by default.
Required changes to how the site template stuff is used to initialize dotlrn (do
NOT EVER hardwire ANYTHING into an after_instantiate callback that MIGHT SOMEDAY
want to be configurable! Sheesh!!!!).  Since Selva used the same layouts as
the default new-portal layouts, we didn't run into this problem with Selva.

But I sure as hell ran into it with Zen.

Localized a couple things in Zen too.